+++ picast/discovery.py
@@ -19,6 +19,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 
 import zeroconf
+from logging import getLogger
+
 from picast.settings import Settings
 
 
@@ -27,12 +29,14 @@ class ServiceDiscovery():
 
     def __init__(self):
         self.config = Settings()
+        self.logger = getLogger(self.config.logger)
         self.zc = zeroconf.Zeroconf()
 
     def register(self):
         service_info = zeroconf.ServiceInfo('_display._tcp.local.', 'PiCast Remote Display._display._tcp.local.',
                                             addresses=[self.config.myaddress.encode('ascii')], port=self.config.rtsp_port)
         self.zc.register_service(service_info, ttl=60, allow_name_change=False)
+        self.logger.info("Register mDNS/SD entry.")
 
     def lookup(self):
         service_info = self.zc.get_service_info('_displaysrc._tcp.local.', None)